Seperated the middleware
